Drasil.Website.About

Description

About Drasil

About Section

wikiSentence :: Sentence

Reusable wiki page sentence.

aboutSec :: Reference -> Reference -> Reference -> Reference -> Reference -> Reference -> Reference -> Reference -> Reference -> Reference -> Reference -> Reference -> Reference -> Section

Creates the about section.

aboutParagraph1 :: Reference -> Reference -> Sentence

Paragraph to about Drasil and its goals.

aboutParagraph2 :: Reference -> Reference -> Reference -> Sentence

Paragraph to describe the layout of the rest of the Drasil website.

aboutParagraph3 :: Sentence

Lists artifacts that Drasil generates

currentlyGeneratedArtifacts :: Contents

aboutParagraph4 :: Sentence

Lists artifacts that we hope to generate with Drasil

futureGeneratedArtifacts :: Contents

aboutParagraph5 :: Reference -> Sentence

Paragraph describing information encoding

aboutParagraph6 :: Reference -> Sentence

Paragraph describing chunks

aboutParagraph7 :: Reference -> Sentence

Paragraph describing recipes

aboutParagraph8 :: Reference -> Sentence

Paragraph describing GOOL

aboutParagraph9 :: Reference -> Reference -> Reference -> Reference -> Sentence

Paragraph providing a link to Drasil papers and documents