Drasil.Website.Body

Description

Gathers and organizes all the information for the Drasil website.

Functions to Generate the Website Through Drasil

printSetting :: FolderLocation -> PrintingInformation

Printing info to get document to generate. Takes in the FolderLocation.

mkWebsite :: FolderLocation -> Document

Instead of being an SRSDecl, this takes the folder locations and generates the document from there.

data FolderLocation

Folder locations based on environment variables (using getEnv in Drasil.Website.Main).

Constructors

Folder

Fields

depL :: FilePath

Deploy location. Currently unused, but may be needed in the future.

docsRt :: FilePath

Haddock documentation root file path. After using make deploy, this should be deploy/docs.

exRt :: FilePath

Example root file path. After using make deploy, this should be deploy/examples.

graphRt :: FilePath

Package dependency graph root file path. After using make deploy, this should be deploy/graphs.

analysisRt :: FilePath

Analysis root file path. After using make deploy, this should be deploy/analysis.

typeGraphFolder :: FilePath

Type graphs root file path. After using make deploy, this should be deploy/analysis/TypeDependencyGraphs.

classInstFolder :: FilePath

Class-instance graphs root file path. After using make deploy, this should be deploy/analysis/DataTable/packagegraphs.

repoRt :: FilePath

Repository root, used for linking to generated code in GitHub.

buildNum :: FilePath

Deploy build number. Currently unused.

buildPth :: FilePath

Deploy build path. Currently unused.

packages :: [String]

List of Drasil packages taken from the Makefile.

si :: FolderLocation -> SystemInformation

System information.

sections :: FolderLocation -> [Section]

Puts all the sections in order. Basically the website version of the SRS declaration.

symbMap :: FolderLocation -> ChunkDB

Needed for references and terms to work.

getSysName :: SystemInformation -> IdeaDict

Helper to get the system name as an IdeaDict from SystemInformation.

usedDB :: ChunkDB

Empty database needed for si to work.

allRefs :: FolderLocation -> [Reference]

Holds all references and links used in the website.

webName :: CI

Used for system name and kind inside of si.

web :: CI

Used for system name and kind inside of si.

Header Section

headerSec :: Section

Header section creator.

imageContent :: LabelledContent

For the drasil tree image on the website.

gitHubRef :: Reference

Used for the repository link.

wikiRef :: Reference

websiteTitle :: String

Hardcoded info for the title, URL, and image path.

gitHubInfoURL :: FilePath

imagePath :: FilePath

Footer Section

footer :: FolderLocation -> Section

Create the footer section.

copyrightInfo :: Sentence

footer contents.