Drasil.Website.CaseStudy

Description

Case Studies table for the different choices available when generating code from Drasil. To be used in the Drasil website.

Case Studies Section

caseStudySec :: Section

Creates the Case Study Section.

caseStudiesTitle :: String

Section title.

caseStudiesDesc :: String

Section description.

legendIntro :: String

Introduce the Case Study Table Legend as a list.

mkCaseTable :: RawContent

Creates the Case Study Table

Manipulating info from Example -> CaseStudy

After taking the information about the examples from Example.hs, convert each example into its own case study.

data CaseStudy

Holds individual case studies. System info may not be needed, but it is still nice to keep around for now.

Constructors

CS

Fields

systemCS :: System

Each case study needs a name, so use system information.

progName :: Sentence

A case study may have different program names for the same example (ex. Projectile).

choicesCS :: Choices

Each case study has code that is generated from a set of choices.

mkCaseStudy :: Example -> [CaseStudy]

Converts a list of examples into a list of CaseStudies. Currently, projectile is the only one that has more than one set of choices, so we take the naming scheme from there.

Display CaseStudy Information as a Table

We first need the helper functions to convert Choices into a displayable format (as a Sentence). Those are defined in the section below to reduce clutter. Then we make the header row, table body, and helper for the table body functions.

headerRow :: [Sentence]

Hardcoded header row for the Case studies table

tableBody :: [CaseStudy] -> [[Sentence]]

Creates the case study table body.

displayCS :: CaseStudy -> [Sentence]

Converts a case study into a table row for easy display.

Case Studies Table Legend

Next, we need the legend to explain the Case Studies Table. These functions are essentially hard-coded and also defined below.

data CSLegend

Each entry for the case studies table legend. The title should be the same as the header.

Constructors

CSL

Fields

ttle :: String

Legend title.

symbAndDefs :: [(String, String)]

Legend symbols along with their respective definitions.

caseStudyLegend :: RawContent

Make the legend for the case study table as a list.

mkLegendListFunc :: CSLegend -> ItemType

Helper to convert the Case Study legends into list items.

mkTandDSent :: (String, String) -> ItemType

Should eventually take Sentences instead of Strings. Converts into the format of "symbol - definition".

caseStudyTitle :: String

Case Study Table column headers.

modularityTitle :: String

Case Study Table column headers.

implementTypeTitle :: String

Case Study Table column headers.

loggingTitle :: String

Case Study Table column headers.

inStructTitle :: String

Case Study Table column headers.

conStructTitle :: String

Case Study Table column headers.

conRepTitle :: String

Case Study Table column headers.

realNumRepTitle :: String

Case Study Table column headers.

legendEntries :: [CSLegend]

Case study legend entries.

modularityLegend :: CSLegend

Modularity or Separation of software.

implementationTypeLegend :: CSLegend

Software implementation type.

loggingLegend :: CSLegend

Compiler logging statements.

inputStrLegend :: CSLegend

Input value structure.

conStrLegend :: CSLegend

Constant value structure.

conRepLegend :: CSLegend

Constant value representation.

realNumRepLegend :: CSLegend

Real number representation.

Helper functions to create the case study table rows.

These functions act like a version of show for each different type of Choices, but tweaked to fit inside a table.

getMod :: Modularity -> Sentence

getImp :: ImplementationType -> Sentence

getLog :: [Logging] -> Sentence

getInstr :: Structure -> Sentence

getConstr :: ConstantStructure -> Sentence

getConRep :: ConstantRepr -> Sentence

getRealNum :: [CodeType] -> Sentence