Drasil.Website.Example

Description

Create the list of Generated Examples for the Drasil website.

Gather Example Information

First gather all information needed to create an example. This includes system information, descriptions, and choices. These will also be exported for use in CaseStudy.hs.

data Example

Each Example gets placed in here.

Constructors

E

Fields

systemE :: System

Example system information. Used to get the system name and abbreviation.

choicesE :: [Choices]

Some examples have generated code with specific choices. They may also have more than one set of choices, so we need a list.

codePath :: FilePath

Generated code path.

srsDoxPath :: FilePath

Generated documents & doxygen path

allExampleSI :: [System]

Records example system information.

allExampleChoices :: [[Choices]]

Records example choices. The order of the list must match up with that in allExampleSI, or the Case Studies Table will be incorrect.

allExamples :: [System] -> [[Choices]] -> FilePath -> FilePath -> [Example]

Combine system info, description, choices, and file paths into one nice package.

examples :: FilePath -> FilePath -> [Example]

Calls allExamples on allExampleSI, allExampleDesc, and allExampleChoices. Can be considered a "default" version of allExamples.

Functions to create the list of examples

fullExList :: FilePath -> FilePath -> RawContent

Create the full list of examples.

allExampleList :: [Example] -> [ItemType]

Create each example point and call individualExList to do the rest.

individualExList :: Example -> [ItemType]

Display the points for generated documents and call versionList to display the code.

versionList :: (Example -> Lang -> String -> Reference) -> Example -> [ItemType]

Takes a function that gets the needed references (either references for the code or doxygen references) and the example to create the list out of. For examples that have more than one version of generated code (more than one set of choices) like Projectile, we generate the code and doxygen references for each.

showLang :: Lang -> String

Show function to display program languages to user.

Examples Section Functions

exampleSec :: FilePath -> FilePath -> Section

Example section function generator. Makes a list of examples and generated artifacts.

exampleTitle :: Sentence

Example section title.

exampleIntro :: Sentence

Example section introduction.

generatedCodeTitle :: String

Example list titles.

generatedCodeDocsTitle :: String

Example list titles.

Helper functions in getting References for SRS, code folders, and Doxygen

convertLang :: Lang -> String

Similar to showLang, but for use within Drasil for Referencing and UIDs.

getCodeRef :: Example -> Lang -> String -> Reference

Generate a reference towards the code folder. Uses getCodePath to find the code path.

buildDrasilExSrcRef :: Example -> Reference

Similar to getCodeRef, but builds the source code references

getDoxRef :: Example -> Lang -> String -> Reference

Similar to getCodeRef, but gets the doxygen references and uses getDoxRef instead.

getSRSRef :: FilePath -> Format -> String -> Reference

Make references for each of the generated SRS files.

getSRSPath :: FilePath -> Format -> String -> FilePath

Get the paths of where each reference exist for SRS files. Some example abbreviations have spaces, so we just filter those out.

getCodePath :: FilePath -> String -> String -> FilePath

Uses repoRt path (codePath in this module).

Get the file paths for generated code and doxygen locations.

getDoxPath :: FilePath -> String -> String -> FilePath

Uses exRt path (srsDoxPath in this module).

Get the file paths for generated code and doxygen locations.

exampleRefs :: FilePath -> FilePath -> [Reference]

Gather all references used in making the Examples section.

getCodeRefDB :: Example -> [Reference]

Helpers to pull code and doxygen references from an example. Creates a reference for every possible choice in every possible language.

getDoxRefDB :: Example -> [Reference]

Helpers to pull code and doxygen references from an example. Creates a reference for every possible choice in every possible language.

getAbrv :: Example -> String

Helper to pull the system name (abbreviation) from an Example.